+++ lib/main.dart
@@ -1,6 +1,5 @@
-
 import 'package:daip/screens/mapDemo.dart';
-import 'package:daip/widgets/botTile.dart';
+import 'package:daip/widgets/home_handler.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
@@ -26,35 +25,6 @@ class MyApp extends StatelessWidget {
   }
 }
 
-class HomeWidget extends StatelessWidget {
-  const HomeWidget({super.key});
 
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        toolbarHeight: 120,
-        title: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 16.0),
-          child: Text("DAIP"),
-        ),
-      ),
-      body: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 16.0),
-        child: GridView.count(
-          crossAxisCount: 2,
-          childAspectRatio: 0.85,
-          children: <Widget>[
-            BotTile(connected: true, img: "images/bot.png",name : "Lorem ipsum something #249 ggez"),
-            BotTile(connected: false, img: "images/bot.png",name : "Lorem ipsum something #249 ggez"),
-            BotTile(connected: false, img: "images/bot.png",name : "Lorem ipsum something #249 ggez"),
-            BotTile(connected: false, img: "images/bot.png",name : "Lorem ipsum something #249 ggez"),
-            BotTile(connected: false, img: "images/bot.png",name : "Lorem ipsum something #249 ggez"),
-          ],
-        ),
-      ),
-    );
-  }
-}
